+++ uniform_cost_search/deno.json
@@ -1,7 +1,7 @@
 {
   "tasks": {
     "dev": "deno run --watch main.ts",
-    "start": "deno run --allow-read main.ts"
+    "start": "deno run --allow-read src/main.ts"
   },
   "imports": {
     "@std/assert": "jsr:@std/assert@1",
